lang.smodel: use ModuleIdentity from lang.modelapi
utilize the fact there's 'extends' dependency between the two to facilitate migration.
'extends' is there to use menu contributions from modelapi, once we have a mechanism to aggregate these without need to extend, we would not be allowed to use 'extends' between concepts, too, but I expect deprecated concepts to vanish well before the new mechanism is in place.