mpsExecutionPlugin: explicit dependency to mps-boot-util
It used to be exposed through mps-kernel java module, but I don't like the idea that every module get mps-boot-util in their classpath, when only few need it.
Alternative fix is to add it to MPS.Core dependencies (as it's the way boot-util classes are visible to Ant_Command), but this effectively gives boot-util jar to every other module, which I'd like to avoid.
Indeed, MPS.Core exposes these classes (CommonPath adds jar, additional.classpath adds respective classes/), and this inconsistency in build scrips is not quite nice, but this helps me to capture structure of dependencies better.
Therefore, I put explicit compile dependency for the only module that can not get it other way (MergeDriverPacker from vcs is compiled as java module and get the dependency naturally).