MPS-30303 http support throws in master
Error introduced in ecb7996b: The problem occurs due to races between
the registration of registryKey extensions of http-support plugin and
the initialization of built-in-server. If latter comes faster, then
MPSInternalPortManager#PORT will be initialized with specified default
in second parameter. Otherwise (what can occasionally happens due to async),
MPSInternalPortManager#PORT will be initialized with a value from
http-support/plugin.xml. There was no value specified, what is treated as
"" by IDEA, that is not a valid int value.