Fixing 'rm -rf <lib root> && git checkout --' for the library root fs update
IDEA sends changed event on the jar!/meta-inf/module.xml files.
Of course we cannot find modules there.

The most logical solution I can come up with is to listen to the
parenting jar.
It seems that anyway it does not make sense to listen to ***/module.jar!/META-INF/module.xml.

Moreover, when we listen to the jar file only, we do not get million of updates for the classfiles
Though MPS does not respond to the **-src.jar updates.
We listen only for deployment, that still needs to be fixed (since it is
the sources we have our models in)