Providing for runtime solutions to work on distribution
Build language still does not support the simplest scenario:
to package the solution with java stubs which are also included as a
java library.

I am doing in the similar way it has been done with
bl.references.runtime and mps-references.jar.
Though later I will support this in the build language.