lang.build: use jarLocations with java module dependencies 4
behavior and constraints rt with JavaModule dep and respective jarLocations layout
To make stubs availabe at runtime, duplicate java_classes model root with a 'yet-to-be-known' jar name (it's build scripts that defines the name).
This approach is faulty not only because of duplciated name knowledge, but also as it doesn't work in MPS-IDEA plugin (there's no platform_lib with mps-*.jar).
However, we are not aware of anyone using stubs in idea plugin at the moment, and it's the simplest possible approach at the moment.