MPS-31163 + somewhat MPS-31106: show implementations also for classifier
reference; merge ShowDefinition and ShowImplementations actions

Check if a subject of show definition is an instance of ClassifierType and in
that case try to apply ShowImplementations to its referent. For that flow, it
seemed natural to merge ShowDefinitition and ShowImplementations, so now we have
less similar-ish actions in a keymap.

Make caption of ShowDefinitionInMenu more specific to contrast it with
ShowDefinition.