Clearing serialization caches of IDEA (similarly to DynamicsPlugins.kt) + trying to clear menus from the actions with disposed cl
The latter is not a big problem (there could be only one generation of outdated action classes in the menus)

Also I have noticed some stale classes in the Disposer's
ObjectTree#myDisposedObjects which is weak on keys not values.
Some #backtraces of Throwables contains our outdated classes.

Moreover SearchEveryWhereUI holds onto the outdated action classes which
is not to be cleared without changing IJ codebase.

I believe we will be able to fix these problems along with advancement of dynamic plugins
in IJ.

Most of these problems have been with us since the old times;
Since they did not affect us that much all this time, I do not consider
them as urgent.