Follow up to the changes in ModelFactory
Note that I left duplicated code in the implementation of MF#canCreate
method.
Probably it makes sense to avoid it by extracting it but I am not sure
which way is the best for it right now.

Also there is a possibility that #supports will be removed (since I
believe that adding the direct knowledge about the mapping between
factories and data sources is wrong)