MPSSPRT-242 Node tabs don't show names returned from concept constraints properties
NodeVirtualFileSystem: try to find virtual file not only by changed node itself (will be null for most of the cases), but also for containing root.
This way platform is notified about virtual file change. As result presentation in UI will be updated and editor tab as one of them.
Additionally propagate reference change that was ignored before.

MPSEditorUtil: return current editor node for all types of NodeEditorComponent. In this case it is not required to fall back to virtual file in com.intellij.openapi.fileEditor.impl.EditorTabPresentationUtil#getEditorTabTitle - file name is derived from this node any way.