MPSSPRT-253 ver. 1.5 -- trying to micromanage PluginLoaderRegistry
PluginLoaderRegistry tries to ensure that the EDT-freezing activity
which takes place in UpdatingTask#run, runs under some kind of progress.
Though most of the time it must work (when IDE is opened and we running
for example rebuild), the app initilization appears to be the harder time when
  there is no progress and no way to get it.

In this change I replace the Timer-based waiting for app components to
load with a listener + invoking UpdatingTask#run with empty progress
when app components are initialized.
I could not find why Task#queue does not work in this case.