MPS-32186 Java stubs are not refreshed
When *jar contents changed, IJ sends CONTENT_CHANGED event on the jar
file itself and thats it (at least on my mac).

While FolderSetDataSource (or any other source) is not subscribed to
updates on the jar path.
It waits for updates of the file inside the jar.
The best solution I found so far is:
1. to fix the problem with merging jar separator in path parts (see my
previous commit or so)
2. employ the listening preferences I developed some time ago in order
for model listener to track the changes of its _ancestors_.
The 'changes' means the event of the type 'CONTENT_CHANGED', and to my
understanding, it never comes from a directory, only from a file
(archives included)

This patch must remove the problem with model sources in the packaged
modules as well.