TreeHighlighter improvements
Group notifications for TreeHighlighter
Refactor CounterMap, removing excessive maps
Propagating messages to ancestors (need to fix for features that are visible only in the tree)
Together with #getParent caching drastically total time improves from 120s to 20s
400 * 20 changes (roots * members)
The time spent in tree highlighter is negligible now (~2 secs);

For this particular scenario the hotspot is again in
NodeFileStatusMapping.

What can be done further:
* there is some dry highlighting (wasting read action time) because we
pull vcs changes but only for the first added node, when others are
added, they see that the model is already pulled and seek for updates in
vcs features holder.
* virtual packages are not with messages. We need to project all coming
features (from vcs) against the tree features we have and store the
status for each node that is present in the tree. All the other features
must be dropped. This will be faster and better from ideological point
of view. Instead of storing both feature tree from UI and feature tree from VCS
changes only one must be left.