MPS-32354 Out-of-process test execution hangs and breaks persistent storage
When FileBasedIndex dispose itself (see FileBasedIndexImpl#performShutdown), it calls ModelInputStream#readModelID and in a case of if-else branch with ModelOutputStream.MODELID_STRING it uses PersistenceFacade.
By that time PersistenceFacade is already disposed and lead to NPE, what corrupts the persistent file storage.

MPSCoreComponents calls that dispose of PersistenceFacade.
In the same time MPSCoreComponents is connection between MPS kernel and IntelliJ platform.

To avoid NPE dispose order need to be maintained:
FileBaseIndexImpl does not implement a Disposable interface itself.
FileBasedIndexImpl.MyShutDownTask is used as disposable to connect FileBasedIndex and ManagingFS (see FileBasedIndexDataInitialization#prepare).
So ManagingFS can be used to handle dispose order and used as a child to register MPSCoreComponents in the Disposer.

---

There is no need to call TestApplicationManager#dispose method any more.