ProjectPane: reliable message propagation
Each updater has to deal with status/messages of nodes, and it's highlighter that knows better which nodes and when to re-calculate.
Keep in mind than namespce nodes can appear anywhere between module and model (not only above modules), it's tricky to guess proper propagation strategy within e.g. visitModel(), it's easier if external code do that for us.