build lang: fix discovery of standalone generator modules in layout
When a standalone (not part of mpl) generator module is mentioned as a dependency (e.g. to address implicit dependency in generator tests by module reference), their layout module could not be properly identified to to legacy pieces in exports(), location() and unpack() methods that didn't account for mpst descriptor files.