build lang: BL_Copy replaced with BL_Files
As I understand, BL_Copy and BL_CustomCopy have been replaced with BL_File/BL_Files; it's not possible to create new instances(empty subst menu).
However, there's a lot of uses in templates, in fact, new constructs are treated as aliases for old one (likely, as it's much more work to refactor these).
If I do plan to refactor, I could make use of intention as it now copies tempalte macros, if any, to a new node.