MPS-32637 EditableSModel#save is explicitly async; adding #save(SaveOptions)
Storage-memory conflicts are resolved using a special resolver instead of listening mechanism.

Using #onConflict is incorrect because the internals of EditableSmodel
will not save anyway unless there is a special listener hanging which
resolves the conflict and force-saves the model.

Secondly, there might be many listeners.
Can there be two resolver among them? Presumably not.
What will happen if we write another platform layer over the current
EditableSModelBase? How do we guess that we need to register a resolver
listener? It appears that this core api is not closed.
I figured that this was an obscure decision and I have moved away from that to having a dedicated resolver.

Fixing MPS-32637 by this explicit asyncronous #save and by protecting the
EditableSModelBase#resolveConflict0 with atomic, which is released only
upon the end of the future (=when user chooses something).