MPS-32683 new module doesn't get imports updated
SilentModuleVersionUpdater rightfully doesn't look into modules that are not part of a project.
However, the moment module is being added to the project, it's not yet treated as 'module of the project'.
New module doesn't get tracked and no version update is performed for it once its model is changed.