MPSSPRT-285 Project migration is not executed in headless migration: cleaning ClassNotFoundException
CNFE is due to the fact that build.pluginSolution contains the build run
configuration which generates into code with ConfigTypeEnvoy which lies
under 'jetbrains.mps.execution.api' module.
This module comes under idea plugin with the same name.
In the migration task there are all modules from the mps distribution
and no plugins available, and since classloading subsystem did not
consider a solution with an invalid idea facet as ineligible for deploy
CNFE happened.

I fix the classloading subsystem problem here.
The whole subsystem is to be replaced: just following the surrounding
logic.