Fixing the GetDataRule problems for MPS-32710 + MPSSPRT-283 [MPSI-21]
I decided to use BaseNodeEditor as a reasonable place with a project.
Its panel is an ancestor of NodeEditorComponent.
panel#getData delegates to NodeEditorComponent#getData (looks strange to
me but who am I).
the project is there, lets return a project.
Notice that NEC#getData if invokeD NEC#getCurrentProject leads to SOE,
because its ancestor delegates all the unknown keys to its descendant.
now when the project is a known key, this will not happen, but this
reversed delegation is something to think about.
Commit history reveals some ancient bug with alt-insert when this
delegation appears.
Probably it is worth reconsidering but I am not too eager to do that.
To summarize, project is always within the BaseNodeEditor, run
configurations have the project and all the other keys at
BaseNodeEditor#MyPanel.
On the other hand MPS-32710 is still ok, since we do not have MPSProject
at any time in LocationRule like we used to have before the artem's fix
of MPS-32710