FrameRule is also fixed in a way
There was a serious potential problem due to the fact that
WindowManager.getInstance().getVisibleFrame() looks for the first
visible frame which is not necessarily the frame on the top.
So the returned frame might be not the one you are looking for.
The fix uses another API which proved working better on my computer at
least.
Adding a warning to get whether we are in potential danger in 203.