MPS-32798 Model those are not tracked in version control may have Diff marks and project pane highlighting
and many more...

Not tracking vcs differences for non-tracked models;
Eager calculation in NodeFileStatusMapping;
Defect with untracked models becoming tracked and many more.

What works in both persistences:
* tracked becoming untracked and vice versa,
* adding/removing a node from the model gets the model modified;

What I know that does not work:
* module is never marked in any way, this part is skipped in the current
  logic of highlighting;
* added model in per-root persistence with untracked root will be show as
  added model and added root;

The problem: the notifications for the model are not there, so all the
cases are handled ad hoc at every listener.
The proper way is to move feature logic (which is already in the
TreeHighlighter) to the level of changes tracking, so that module/model
changes tracking does not differ from the node feature tracking logic.
Otherwise it will be unbearable.
Right now since we are in bugfix as always I have to add
the method #modelStatusUpdated just to get some response for a tracked
model becoming untracked (otherwise nothing happens since in the empty
model there is no changes, changeset is always empty).

What to do here further:
Fully accessible tree of features/changes (or wrappers) over
Smodule->Smodel->Snode hierarchy.
The status calculation being provided at some common
place; TreeHighlighter and NodeFileStatusMapper being the clients.