MPS-33186 Сolors in Annotate
The colors for the "Order" color option were calculated using the total number of model commits.
It could happen that only small part of the commits had changes for the annotated root.
In this case the calculated colors would differ very little.
Now, the colors are calculated only for those revisions where the root was changed.