MPSSPRT-319 Fixing the generation is not available while indices are building notification
The first part is to make the saving part of the make action synchronous, that's what I did.
This is not important in the scope of the current issue, since we were in EDT in this
specific case, so the save was done synchronously.
It is the response of the vcs subsystem which
invoked some kind of async indexer if more than 20 files are changed on the disk.

The second part:
In order to reproduce one has to have quite a large per-root model so that the
number of files, written on save, is at least 20 (see
ChangedFilesCollector#ensureUpToDateAsync).
We make a non-subtle if on where we are in edt, in order to flush dumb
tasks queue.
It works on the customer's per-root model.

In the ideal world make is running in a background read, and then DumbService
has the method #runReadActionInSmartMode for this.
What happens here is too complicated and I do not want to take risk and
make any significant changes.