MPS-33231 blinking diff marker in editor gutter.
The marker is blinking when the added node is then deleted. The reason is the following:
there is a ChangeEditorMessage instance for the added node in EditorHighlighter. When we
delete the node it's parent is set to null and the editor is repainted. However, at this
point the message is still kept in the EditorHighlighter. The node with parent==null is
considered as undirect root and thus the first line is marked for the ChangeEditorMessage
of deleted node.

Previous fix 8498aa1d0ed362d9915c64076af79e080e33b0ec with extra model listener is reverted here.
Instead, the messages provided by EditorHighlighter are filtered.