TestMakeUtil & WorkerHelper use 2-step ModuleMaker
Use forceCompile=true in tests as generally we care to compile code in tests, not to check 'is dirty' logic.

In cmdline tools, forceCompile=true is not so obvious (e.g. if we want to avoid re-compilation of existing code when running cmdline build again and again. There's always Ant/Gradle/OS mechanism to ensure clean build, if necessary). However, decided to go with forced compile as seems cheap compared to investigation of stale builds (i.e. prefer to make every thing possible to ensure artifacts are up-to-date).