MPS-33608 Failed to diff models from Annotate.
Models for the revisions to compare are loaded now each time when 'Change to Editor Tab' or
'Change to Separate Window' button is pressed. Model loading is performed now as
a background task, similar to usual 'Show diff' action. Some refactoring is done in order to
reuse existing code for 'Show diff' action.