ModuleRuntime: provisional support for activators
to address the fact one can't use lang.plugin.standalone to contribute anything to the platform from a module that is part of mpsBootstrapCore build project. lang.plugin.standalone lives in mpsWorkbench (along with its MPS.Workbench RT), ant it's not an option neither to move workbench code into mpsBootstrapCore, not to move make.facets out of it. Hence, provisional mechanism to get some code from a solution activated on module load.