Refactor JavaDebugEvaluate
Instead of custom facet with in-memory compilation and output resource with a classloader, transform model into temp location much like console does, using MPS's native compile & reload make facets.
Even if in-memory compilation is faster, it's not enough to justify complexity of the code. If performance is an issue,
we may further refactor to produce textgen-in-memory objects and then compile these. To get to the point of in-memory output, however, I need to refactor other places with TextGenOutcomeResource, and one more in JDE complicated that a lot.