Changing hard assertions of path format to warnings
PathFormatChecker defines two methods that throw exception if a
non-system path delimiter is found in a path.
Problem is, on *nix systems path may contain a backslash.
Such paths cause this exception to be triggered, which is undesired.
Instead we trigger just a warning.
In order to be able to work properly with such paths, we need to finish
our job with Path classes which store not just a string, but path parts.
Then it will become possible to extract a canonical representation for
such objects.

fisakov & apyshkin