MPS-33445 correct diff when some models are not loaded.
If the model is not loaded then the null value is handled in all necessary places
and the label 'Failed to load model' is put to the DiffEditor. Similarly,
null values are used for corresponding ChangeSet and ChangeGroupLayout instances.

DiffEditor constructor got two parameters 'model' and 'node id' instead of a
single 'node' parameter. This allows one to distinguish the cases when model is not
loaded and when the root is not found in the loaded model.

NextPreviousTraverser is refactored a bit.

A bug with not showing component is fixed in TripleChangeGroupLayout#paintPolygons.

ModelDiffViewer and ModelDifferenceViewer are refactored a lot. Handling of conflicts in three
model case is completely moved from ModelDifferenceViewer to ThreeSideRootDifferencePane.