lang.descriptor: per-root aspect mechanism
To generate instantiation of aspect class bound to node, not an aspect model, introduced a new switch intended for extensions with a rule to match node's concept
Make use of this new mechanism in customAspect.documentation sample.
Qualifified apsect class name is sort of a hack, proper way is to declare GP for the aspect which synchonizes with a common CP with lang.descriptor.