MPS-34059 compile fails from a forked process
Force use of UTF-8 encoding as it used to be with eclipse compiler (see JavaFile.getContent() that used to force UTF-8 for any java source file, supplied to the compiler as char sequence).
Alternative fix would be explicit 'file.encoding' for a forked process in MpsLoadTask, I just feel this one is safe for bugfix.