Move df.presentation out of [mps-workbench]
Actions from df.pluginSolution and bl.pluginSolution visualize DF with IGraphCreator. No reason for this code to be part of [mps-workbench]. Provided bl.pluginSolution already depends on df.pluginSolution, I decided latter is proper place for the presentation code (i.e not a separate solution). Anyway, j.m.ide is not the proper home for the DF presentation code.