MPS-34141 enable Favorites tool.
In the platform Favorites tool functionality was merged into Bookmarks,
see IDEA-176014. By default Favorites were switched off with the
registry key 'ide.favorites.tool.window.applicable'.

It makes sense to make it on by default in MPS since MPS's bookmarks are
not as good as those in the platform.