Set/Remove VF action for a module to exclude nested modules
These actions used to find out if module is nested under another module with project.getPath() == null assumption.
As it's no longer true, need to come up with an alternative approach.
Using TreeNodes seems right as we reveal virtual folders in ProjectPane only, and it doesn't seem right to manage VFs where SModule is present but it's not a project pane.