MPS-33986 lang.build better message for duplicated modules
Generally, we don't care about module name, it's id that's relevant.
However, generally we don't encourage same names as it brings confusion and may result in build defects, as we use module name for file names.
Report duplicates as warnings instead of error as it's not that severe, and tell 'same project' case from 'different build projects' to give meaningful message.