MPS-34925 detect PC to unload the moment module instance is *almost* valid
relates to merged MPSSPRT-413 changes. Now, we use JMF to detect if module comes with extensions, but CLM notification about modules unloaded comes the moment modules are already detached from a repo, and all facets gone.
As mentioned in MPS-34925, eventually we shall use LanguageRegistry/ModuleRuntime notifications, not to rely on SRepository and SModule lifecycle.