Temporarily disable logging errors for captured stderr
Normally any stderr line captured from forked process should be logged
as error, with the exception when received while a test is being run,
in which case it should be reported as test stderr via service message.

However, when lots of data is pushed to stderr (from threads other than the
one running the test), a "test finished" message might appear in the
stream earlier than the rest of the data, causing TeamCity to recognized
a test as failed.