Module properties: model roots to show actual values
ModelRootDescriptor is pure persistence holder. Recently, ModelRoot changed to complete initialisation the moment it has SModule attached (to resolve ${module} macro properly), not the moment it receives persistent configuration. Editing of blank ModelRoot instances no longer works and we have to grab actual, initialized ModelRoot instances. For MPS-as-IDEA-plugin it constitutes an issue as there's no SModule instance then, need to address this separately.