build lang: use SModule/SRepo all the time + JMF for java libs
With SModule being processed during full import, there's little justification to stick to MD (non SModule) info at check time. On a contrary, seems right to use SRepo/SModule for check and partial import scenarios, and not to use for full-import, provided the idea was to keep build project full of information enough to perform a build.
However, I'm not sure if we ever get to the point of not using SModule during 'full import' (or, for that matter, to use build project data only during generate, w/o need to import anything in addition).