MPS-35612 and MPS-27464 "Method impl could not be found" exception
This is an alternative for 67b4483b (MPS-23860), as the warning in structure models is quite annoying and hard to deal with (forcing user to write implementations for abstract methods is not very enjoyable activity).
As long as it's MPS-by-design approach, let MPS handle it instead.
I left the message for now, just lowered its severity to 'info' as it still worth to let LD know there's something not entirely fine with the declaration. Well, I struggle to find better word instead of 'fine' as in my opinion, it's perfectly ok to have [1] aggregation of an abstract child.
And FTR, I'm not quite satisfied with the try-catch approach, just find it tedious to refactor all the code around virtual method invocation.