Use userObject of GenTask/GenStatus to pass dictionary around
TextGen facet doesn't have access to MResource and, in order to access
the required flag cleanMake, alternative ways have to be explored. The flag
cleanMake is per module and specified in MakeKeys.CLEAN_MAKE, accessible via MResource.
As MResource is now a Dictionary, we can simply use that as the user object.