Limit Solution@descriptor to non-empty models
With use of ActivatorUnit instances, it's only lang.ext.plug solution at the moment that got non-empty @descriptor model.
The problem is scenario when there's a non-empty model with name matching that of module (and that of @descriptor model), once module is rebuild, we've got output of two models directed into single location.

Since even non-generated @descriptor model answers with hash, have to refresh solutions with LoadExtensions=on