ModulePlugins: lookup module in all visible plugins instead of those declared in same file
As faced in MPS-36475 with mpsVcs exporting a module defined in mpsVcsPlatform, modules may be used in a plugin in a different root, which would cause incorrect detection and ModulePlugins to mark these modules as standalone. On the build script side, a <library> for these modules would be added, which potentially could cause classloading issues.

This change makes ModulePlugins uses all plugins visible from the current file (all dependencies) to look-up a module, instead of checking only the root.