MPS-37203 report java compilation errors for module cycles
When reported individually by module, chances are we get error due to missing package before we get to actual syntax/code problem. There used to be a need to report errors per module, but as it's not required any more, we are free to report errors right from the cycle.