Fix deadlock caused by FakeProcessHandler; MPS-37676
The overridden implementation of BaseOutputReader in FakeProcessHanlder
causes tests to randomly deadlock. MPS-37676 has the complete thread dump
of a deadlocked process.

Presumably, this happens due to com.intellij.util.io.BaseDataReader#waitFor()
waiting for the "error stream" reader to terminate first, which never
happens since the input buffer is empty. We get into this situation because
the process handler receives a premature "process terminated" event caused by,
presumably, InterruptedException in the WaitFor implementation.

To resolve a potential deadlock, we also tell the writer to close in stop() method
of the overridden reader. That method is called before a wait on the reader's future,
so the latter should not block.

See also MPS-24594 and a0e01de948a76fe0f6f1dfaa81a6cb4032287776.