MPS-29589 - detect case-sensitive rename on a case-insensitive FS and in such cases: a) do not log an error and instead proceed to the module dir and module descriptor rename b) in IdeaFile.rename1() proceed with such renames
On case-sensitive FS the condition for getPath() being the same ensures that files differing only in cases will be considerred different files.
On case-insensitive FS the condition for getPath() being the same ensures that two instances of IFile differing only in cases and pointing to the same physical file will be detected as case-only renames.