Fix Kotlin class loading from stubs
The new way of loading metadata from .knm results in classes and class declarations
being iterated in a different order than before. A procedure in ClassRoot was
relying on nested classes being listed before their container. Also, a typo fixed,
which prevented enum memebers from being populated.

The order of declarations within a class has also changed, and is now in accord
with the order reported by IDEA's decompiler. So, the test data had to be fixed, too.