JavaToMps test: drop odd nonStatic patch code
it's odd to modify 'expected' model to get tests green.
Instead, express 'expected' model in a way we get from JavaToMpsConverter.
It's tough situation here, as 'static' keyword is not necessary for nested interfaces,
however, MPS BL editor uses nonStatic as pure AST-kind attribute (i.e. if it's present in the code).
I suppose JavaToMpsConverter shall not consider visual presentation, and shall rather use this property to express semantics.
Alternatively, as long as it's source code, we may want to respect actual keyword in the source file, however, not sure if Eclipse parser gives one to us.