Avoid using transient models as folders in project view; MPS-38463
Transient models may build hierarchy of virtual folders, which correspond
to model names without the stereotype. As a result, some models may be appear
several times in the view. This patch fixes that by making sure the virtual
folder name corresponding to a transient model does include the stereotype,
which prevents the behaviour described above.