Ensure error strip has markers only on leafs/unexpanded branches; MPS-38495
Introduce ProblemHierarchyNode marker interface for nodes that serve dual
purpose: as well as being containers for other objects that can be
annotated with error messages, these can also be targets for their own errors.

The volatile field LogicalProjectViewNode.myHasOwnProblems is updated within
postProcess() method call in order to cache the info that is needed
later in the painting procedure.