GeneratedClass annotation: do not keep full node-ptr
It's a waste of .class space to keep both node-ptr along with omnipresent model-ptr in the annotation.
As long as it's CLASS retention, it's kept in class files and due to MPS approach to load stubs (ignoring Retention) in memory, too.
Decided to keep 'model' + 'nodeId', instead of relying on 'node' and dropping 'model' altogether is the generation logic that
 'moedl' is always present, while 'node' is only when there's original node. Besides, 'model' constants can be shared in class pool, while
 complete node-ptr is less likely to get shared.