Fix error detection and error texts of JUnitRunTypes; MPS-39139
Error in JUnit run configuration now clearly states
`Module %s must be compiled by MPS. Can't run tests.` if the
compiler settings are incompatible with JUnit runner.

Additionally, if the run configuration refers to a test that was
deleted in the model, a "Broken reference[%s]" is displayed
in the test list. This fixes MPS-39141 as well.