MPS-39248 check actual concept kind when reading a model
For a non-migrated model, there could be "ng" flag for INamedConcept.name (instead of proper "ngI").
If it's a first time MAF sees the property, it caches SPropertyAdapter instance with wrong owner
 (SConceptAdapter instead of SInterfaceConceptAdapter), and later can't match BaseConstraintsDescriptor
 (SInterfaceConceptAdapter.equals() doesn't take anything but SInterfaceConceptAdapter).
Models usually get loaded with RegularMetaModelInfo, which is capable to look into actual CDs and hint proper concept kind.